Fix test setup

+++ community/cypher/runtime-spec-suite/src/test/scala/org/neo4j/cypher/internal/runtime/spec/tests/ProvidedOrderTestBase.scala
@@ -1013,7 +1013,7 @@ trait NonParallelProvidedOrderTestBase[CONTEXT <: RuntimeContext] {
   }
 
   test("semi apply should keep order of lhs") {
-    val input = Range(0, sizeHint).map(_ => randomValues.nextInt())
+    val input = Range(0, sizeHint).map(_ => randomValues.nextDouble())
 
     val logicalQuery = new LogicalQueryBuilder(this)
       .produceResults("x")
@@ -1031,7 +1031,7 @@ trait NonParallelProvidedOrderTestBase[CONTEXT <: RuntimeContext] {
   }
 
   test("apply with semi apply on the rhs should keep order of lhs") {
-    val input = Range(0, sizeHint).map(_ => randomValues.nextInt())
+    val input = Range(0, sizeHint).map(_ => randomValues.nextDouble())
 
     val logicalQuery = new LogicalQueryBuilder(this)
       .produceResults("x")
